up roletop

+++ RoleTopMVC/Models/Cliente.cs
@@ -7,8 +7,8 @@ namespace RoleTopMVC.Models
         public string Email {get;set;}
         public string Senha {get;set;}
         public string Nome {get;set;}
-        public int CPF {get;set;}
-        public int Telefone {get;set;}
+        public ulong CPF {get;set;}
+        public ulong Telefone {get;set;}
         public uint TipoUsuario {get;set;}
 
 
@@ -17,7 +17,7 @@ namespace RoleTopMVC.Models
 
         }
 
-        public Cliente (string email, string senha, string nome, int cpf, int telefone)
+        public Cliente (string email, string senha, string nome, ulong cpf, ulong telefone)
         {
             this.Email = email;
             this.Senha = senha;
